Skin tone is pinnable; 21 skins, 24 accessories, 17 mouths; pins move to uint64
Human tones (common, uncommon) can be pinned like any other pin; fantasy
tones are rare or legendary and come only from luck. Pins are one byte per
key in a uint64: background, top, eyes, hair, skin, three spare.

+++ contracts/src/IFaceRenderer.sol
@@ -2,13 +2,13 @@
 pragma solidity ^0.8.28;
 
 /// @notice Draws a face from its seed, its pins and its 1/1 index (255 for none).
-/// `pins` is one byte per pinnable slot, high byte first, 255 for no pin.
+/// `pins` is one byte per pin key, high byte first: the pinnable slots, then the skin, then spare bytes; 255 for no pin.
 interface IFaceRenderer {
-    function tokenURI(uint256 tokenId, uint64 seed, uint32 pins, uint8 one) external view returns (string memory);
-    function svg(uint64 seed, uint32 pins, uint8 one) external view returns (string memory);
+    function tokenURI(uint256 tokenId, uint64 seed, uint64 pins, uint8 one) external view returns (string memory);
+    function svg(uint64 seed, uint64 pins, uint8 one) external view returns (string memory);
     /// @return true when every set pin names a common or uncommon item of a pinnable slot
-    function pinsOk(uint32 pins) external view returns (bool);
-    function pinCount(uint32 pins) external pure returns (uint256);
+    function pinsOk(uint64 pins) external view returns (bool);
+    function pinCount(uint64 pins) external pure returns (uint256);
     function oneOfOneCount() external view returns (uint256);
     /// @return the 1/1 pool index a seed takes with odds poolLength / tokensLeft, or 255
     function luckyFor(uint64 seed, uint256 poolLength, uint256 tokensLeft) external view returns (uint256);
